requirements.txt
langchain==0.1.13
langchain-community==0.0.29
faiss-cpu==1.8.0 
sentence-transformers==2.7.0
numpy==1.26.4
groq==0.9.0
python-dotenv==1.0.1
flask==3.0.3
beautifulsoup4==4.12.3 